Refactor: Convert FavoriteMovies as a component
